SharedMemory.write: accept keyword arguments
I've been using a lot of keyword arguments in some semaphore example code
that I'm working on. It makes the initialization of SharedMemory and Semaphore
objects easier to understand, and I was using the same style in my calls to
SharedMemory.read/write. The write function wasn't written to parse keyword
arguments, though.

The "args_format" char array was already initialized correctly for both Python
2 and 3, so:

* Add another PyObject* to the signature of SharedMemory_write
* Add a "char *keyword_list[ ]" in the same style as the one in
  SharedMemory_read
* Parse arguments with PyArg_ParseTupleAndKeywords instead of
  PyArg_ParseTuple, using the new keyword_list and previously-unused
  args_format

Signed-off-by: Matthew Ruffalo <matthew.ruffalo@case.edu>